requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
httpx
pillow
pytesseract
pdfplumber
PyMuPDF
moviepy
youtube-transcript-api
beautifulsoup4
sentence-transformers
transformers
torch
torchaudio
whisper
faiss-cpu
qdrant-client
numpy
pandas
tqdm
langdetect
sentencepiece
opencv-python-headless
streamlit
python-dotenv
